requirements.txt
aiohttp 
aiosignal 
amqp 
annotated-types 
anyio 
asgiref 
astor 
asttokens 
async-timeout 
attrs 
audioread 
av 
billiard 
blinker 
boto3 
botocore 
Brotli 
cachetools 
celery 
certifi 
cffi 
charset-normalizer 
click 
click-didyoumean 
click-plugins 
click-repl 
colorama 
coloredlogs 
contourpy 
cryptography 
ctranslate2 
cycler 
dataclasses-json 
decorator 
defusedxml 
distlib 
dj-rest-auth 
Django 
django-allauth 
django-cors-headers 
django-debug-toolbar 
django-filter 
django-rest-auth 
django-rq 
django-templated-mail 
djangorestframework 
djangorestframework-simplejwt 
djoser 
dnspython 
drf-spectacular 
duckdb 
einops 
email-validator 
encodec 
executing 
Faker 
faster-whisper 
ffmpeg 
ffprobe 
filelock 
Flask 
Flask-Cors 
flask-mongoengine 
Flask-WTF 
flatbuffers 
fonttools 
frozenlist 
fsspec 
funcy 
google-auth 
huggingface-hub 
humanfriendly 
idna 
inflection 
ipython 
itsdangerous 
jedi 
Jinja2 
jmespath 
joblib 
jsonpatch 
jsonpointer 
jsonschema 
jsonschema-specifications 
kiwisolver 
kombu 
langchain 
langsmith 
lazy_loader 
librosa 
llvmlite 
MarkupSafe 
marshmallow 
matplotlib 
matplotlib-inline 
mongoengine 
mpmath 
msgpack 
multidict 
mutagen 
mypy-extensions 
networkx 
nltk 
numba 
numpy 
oauthlib 
onnxruntime 
openai 
packaging 
pandas 
pandasai 
parso 
pexpect 
pillow 
platformdirs 
pooch 
prompt-toolkit 
protobuf 
psycopg2-binary 
ptyprocess 
pure-eval 
pyasn1 
pyasn1-modules 
pycparser 
pycryptodomex 
pydantic 
pydantic_core 
pydub 
Pygments 
PyJWT 
pymongo 
pyparsing 
python-dateutil 
python-dotenv 
python3-openid 
pytube 
pytz 
PyYAML 
redis 
referencing 
regex 
requests 
requests-oauthlib 
rpds-py 
rq 
rsa 
s3transfer 
safetensors 
scikit-learn 
scipy 
sentence-transformers 
sentencepiece 
six 
sniffio 
social-auth-app-django 
social-auth-core 
soundfile 
soxr 
SQLAlchemy 
sqlparse 
stack-data 
suno-bark @ git+https://github.com/suno-ai/bark.git@773624d26db84278a55aacae9a16d7b25fbccab8
sympy 
tenacity 
threadpoolctl 
tokenizers 
torch 
torchaudio 
torchvision 
tqdm 
traitlets 
transformers 
typing-inspect 
typing_extensions 
tzdata 
uritemplate 
urllib3 
vine 
virtualenv 
wcwidth 
websockets 
Werkzeug 
WTForms 
yarl 
youtube-dl 
youtube-transcript-api 
yt-dlp 
